+++ Mods ChoGGi/Command Center Distance Sort/metadata.lua
@@ -1,0 +1,16 @@
+return PlaceObj("ModDef", {
+	"title", "Command Center Distance Sort v0.1",
+	"version", 1,
+	"saved", 1550145600,
+	"image", "Preview.png",
+	"id", "ChoGGi_CommandCenterDistanceSort",
+	"steam_id", "1655735750",
+	"pops_any_uuid", "820b006e-9e01-4bd0-a706-0917008262ac",
+	"author", "ChoGGi",
+	"lua_revision", LuaRevision,
+	"code", {
+		"Code/Script.lua",
+	},
+	"description", [[On load and every new Sol this will sort the cc list of each building by distance.
+Hopefully makes the drones use the nearest first.]],
+})